requirements.txt
apache-airflow-providers-google
google-cloud
requests